+++ charmcraft/commands/build.py
@@ -32,7 +32,6 @@ from charmcraft.config import Base, BasesConfiguration
 from charmcraft.manifest import create_manifest
 from charmcraft.metadata import parse_metadata_yaml
 from charmcraft.parts import Step
-from charmcraft.providers import get_provider
 from charmcraft.providers.providers import (
     BASE_CHANNEL_TO_PROVIDER_BASE,
     capture_logs_from_instance,
@@ -40,6 +39,7 @@ from charmcraft.providers.providers import (
     ensure_provider_is_available,
     get_base_configuration,
     get_instance_name,
+    get_provider,
 )
 from charmcraft.utils import get_host_architecture
