+++ charmcraft/providers/__init__.py
@@ -16,7 +16,6 @@
 
 """Provider support."""
 
-from ._get_provider import get_provider  # noqa: F401
 from ._lxd import LXDProvider  # noqa: F401
 from ._multipass import MultipassProvider  # noqa: F401
 from ._provider import Provider  # noqa: F401
